First step in moving logic and dependencies into their own section
Big changes are the Organizer and Scheduler
The WebClient Pool and Crawl/Process threads have moved into their own
classes (and moved out of SiteCrawler).

The pom.xml got explicit logging imports (and we have a logback-test.xml
now)
