+++ src/main/java/com/salesforce/webdev/sitecrawler/scheduler/Organizer.java
@@ -1,0 +1,14 @@
+package com.salesforce.webdev.sitecrawler.scheduler;
+
+public class Organizer {
+
+    private final Scheduler scheduler;
+
+    public Organizer(Scheduler scheduler) {
+        this.scheduler = scheduler;
+    }
+
+    public void submitUrl(String url) {
+        scheduler.offerUrl(url);
+    }
+}
