+++ src/main/java/com/salesforce/webdev/sitecrawler/scheduler/Scheduler.java
@@ -1,0 +1,12 @@
+package com.salesforce.webdev.sitecrawler.scheduler;
+
+public interface Scheduler {
+
+    void init();
+    void setId(String id);
+    String getId();
+    void pause();
+    void unpause();
+    void shutdown();
+    void offerUrl(String url);
+}
